+++ e2e-tests/tests/expo.test.ts
@@ -30,10 +30,17 @@ describe('Expo', () => {
       [KEYS.DOWN, KEYS.DOWN, KEYS.ENTER],
       'Do you want to enable Session Replay to help debug issues? (See https://docs.sentry.io/platforms/react-native/session-replay/)',
     ));
-    const testEventPrompted =
+    const feedbackWidgetPrompted =
     sessionReplayPrompted &&
       (await wizardInstance.sendStdinAndWaitForOutput(
         // Enable session replay
+        [KEYS.ENTER],
+        'Do you want to enable the Feedback Widget to collect feedback from your users? (See https://docs.sentry.io/platforms/react-native/user-feedback/)',
+      ));
+    const testEventPrompted =
+    feedbackWidgetPrompted &&
+      (await wizardInstance.sendStdinAndWaitForOutput(
+        // Enable feedback widget
         [KEYS.ENTER],
         'Have you successfully sent a test event?',
       ));
@@ -66,7 +73,7 @@ Sentry.init({
   // Configure Session Replay
   replaysSessionSampleRate: 0.1,
   replaysOnErrorSampleRate: 1,
-  integrations: [Sentry.mobileReplayIntegration()],
+  integrations: [Sentry.mobileReplayIntegration(), Sentry.feedbackIntegration()],
 
   // uncomment the line below to enable Spotlight (https://spotlightjs.com)
   // spotlight: __DEV__,